refactor: ProductDao를 도입하여 다른 레포지토리와 동일한 추상화 단계 구현
