+++ app/src/main/java/woowacourse/shopping/data/repository/product/ProductRepositoryImpl.kt
@@ -1,52 +1,18 @@
 package woowacourse.shopping.data.repository.product
 
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import kotlinx.serialization.json.Json
-import okhttp3.HttpUrl
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import woowacourse.shopping.data.network.ProductDto
+import woowacourse.shopping.data.network.product.ProductDao
 import woowacourse.shopping.domain.Product
 
 class ProductRepositoryImpl(
-    private val client: OkHttpClient,
-    private val baseUrl: HttpUrl,
-    private val json: Json,
+    private val productDao: ProductDao,
 ) : ProductRepository {
-
     override suspend fun loadProducts(
         startIndex: Int,
         pageSize: Int,
-    ): List<Product> = withContext(Dispatchers.IO) {
-        val url = baseUrl.newBuilder()
-            .addPathSegment("products")
-            .addQueryParameter("startIndex", startIndex.toString())
-            .addQueryParameter("pageSize", pageSize.toString())
-            .build()
-        val request = Request.Builder().url(url).build()
+    ): List<Product> = productDao.findAllProduct(
+        startIndex = startIndex,
+        pageSize = pageSize,
+    )
 
-        client.newCall(request).execute().use { response ->
-            check(response.isSuccessful) { "products 요청 실패: ${response.code}" }
-
-            val body = response.body?.string().orEmpty()
-            json.decodeFromString<List<ProductDto>>(body)
-                .map(ProductDto::toDomain)
-        }
-    }
-
-    override suspend fun getProduct(id: String): Product = withContext(Dispatchers.IO) {
-        val url = baseUrl.newBuilder()
-            .addPathSegment("product")
-            .addQueryParameter("id", id)
-            .build()
-        val request = Request.Builder().url(url).build()
-
-        client.newCall(request).execute().use { response ->
-            check(response.isSuccessful) { "products 요청 실패: ${response.code}" }
-
-            val body = response.body?.string().orEmpty()
-            json.decodeFromString<ProductDto>(body).toDomain()
-        }
-    }
+    override suspend fun getProduct(id: String): Product = productDao.findById(id = id)
 }
